improved doxygen documentation
