+++ include/meta/meta.hpp
@@ -23,16 +23,86 @@
 ///
 /// A tiny metaprogramming library
 
+/// \defgroup metafunction Metafunction
+/// Metafunction invocation/composition.
+/// \ingroup meta
+
+/// \defgroup invocation Invocation
+/// Metafunction invocation
+/// \ingroup metafunction
+
+/// \defgroup composition Composition
+/// Metafunction composition
+/// \ingroup metafunction
+
+/// \defgroup logical Logical
+/// Logical operations
+/// \ingroup meta
+
+/// \defgroup algorithm Algorithms
+/// Algorithms.
+/// \ingroup meta
+
+/// \defgroup query Query/Search
+/// Query and search algorithms
+/// \ingroup algorithm
+
+/// \defgroup transformation Transformation
+/// Transformation algorithms
+/// \ingroup algorithm
+
+/// \defgroup runtime Runtime
+/// Runtime algorithms
+/// \ingroup algorithm
+
+/// \defgroup datatype Datatype
+/// Datatypes.
+/// \ingroup meta
+
+/// \defgroup list List
+/// \ingroup datatype
+
+/// \defgroup integral Integral constant
+/// \ingroup datatype
+
+/// \defgroup math Math
+/// Integral constant arithmetic.
+/// \ingroup meta
+
+/// \defgroup lazy_metafunction lazy
+/// \ingroup metafunction
+
+/// \defgroup lazy_invocation lazy
+/// \ingroup invocation
+
+/// \defgroup lazy_composition lazy
+/// \ingroup composition
+
+/// \defgroup lazy_logical lazy
+/// \ingroup logical
+
+/// \defgroup lazy_query lazy
+/// \ingroup query
+
+/// \defgroup lazy_transformation lazy
+/// \ingroup transformation
+
+/// \defgroup lazy_list lazy
+/// \ingroup list
+
+/// \defgroup lazy_math lazy
+/// \ingroup math
+
 /// Tiny metaprogramming library
 namespace meta
 {
+    /// \cond
     inline namespace v1
     {
+        /// \endcond
+
         template <template <typename...> class C, typename... Ts>
         struct defer;
-
-        /// \addtogroup meta
-        /// @{
 
         namespace detail
         {
@@ -51,29 +121,38 @@ namespace meta
         struct list;
 
         /// An empty type.
+        /// \ingroup datatype
         struct nil_
         {
         };
 
         /// "Evaluate" the metafunction \p T by returning the nested \c T::type
         /// alias.
+        /// \ingroup invocation
         template <typename T>
         using eval = typename T::type;
 
         /// Evaluate the Metafunction Class \p F with the arguments \p Args.
+        /// \ingroup invocation
         template <typename F, typename... Args>
         using apply = typename F::template apply<Args...>;
 
+        /// Lazy versions of meta actions
         namespace lazy
         {
+            /// \sa `meta::eval`
+            /// \ingroup lazy_invocation
             template <typename T>
             using eval = defer<eval, T>;
 
+            /// \sa `meta::apply`
+            /// \ingroup lazy_invocation
             template <typename F, typename... Args>
             using apply = defer<apply, F, Args...>;
         }
 
         /// A Metafunction Class that always returns \p T.
+        /// \ingroup metafunction
         template <typename T>
         struct always
         {
@@ -93,11 +172,14 @@ namespace meta
         };
 
         /// An alias for `void`.
+        /// \ingroup metafunction
         template <typename... Ts>
         using void_ = apply<always<void>, Ts...>;
 
         namespace lazy
         {
+            /// \sa `meta::always`
+            /// \ingroup lazy_metafunction
             template <typename T>
             using always = defer<always, T>;
         }
@@ -132,6 +214,7 @@ namespace meta
 
         /// An alias for `std::true_type` if `T::type` exists and names a
         /// type; otherwise, it's an alias for `std::false_type`.
+        /// \ingroup metafunction
         template <typename T>
         using has_type = eval<detail::has_type_<T>>;
 
@@ -149,18 +232,22 @@ namespace meta
         /// template<typename List>
         /// using reverse = foldr<List, list<>, lambda<_a, _b, defer<push_back, _a, _b>>>;
         /// \endcode
+        ///
+        /// \ingroup invocation
         template <template <typename...> class C, typename... Ts>
         struct defer : detail::defer_<C, list<Ts...>>
         {
         };
 
         /// An integral constant wrapper for \c std::size_t.
+        /// \ingroup integral
         template <std::size_t N>
         using size_t = std::integral_constant<std::size_t, N>;
 
         /// A metafunction that computes the size of the type \p T.
         /// \par Complexity
         /// \f$ O(1) \f$.
+        /// \ingroup metafunction
         template <class T>
         using sizeof_ = meta::size_t<sizeof(T)>;
 
@@ -168,23 +255,30 @@ namespace meta
         /// any instance of the type \p T.
         /// \par Complexity
         /// \f$ O(1) \f$.
+        /// \ingroup metafunction
         template <class T>
         using alignof_ = meta::size_t<alignof(T)>;
 
         namespace lazy
         {
+            /// \sa `meta::sizeof_`
+            /// \ingroup lazy_metafunction
             template <typename T>
             using sizeof_ = defer<sizeof_, T>;
 
+            /// \sa `meta::alignof_`
+            /// \ingroup lazy_metafunction
             template <typename T>
             using alignof_ = defer<alignof_, T>;
         }
 
         /// An integral constant wrapper for \c bool.
+        /// \ingroup integral
         template <bool B>
         using bool_ = std::integral_constant<bool, B>;
 
         /// A metafunction that always returns its argument \p T.
+        /// \ingroup metafunction
         template <typename T>
         struct id
         {
@@ -192,17 +286,21 @@ namespace meta
         };
 
         /// A metafunction that is type \p T.
+        /// \ingroup metafunction
         template <typename T>
         using id_t = eval<id<T>>;
 
         namespace lazy
         {
+            /// \sa `meta::id`
+            /// \ingroup lazy_metafunction
             template <typename T>
             using id = defer<id, T>;
         }
 
         /// Turn a class template or alias template \p C into a
         /// Metafunction Class.
+        /// \ingroup composition
         template <template <typename...> class C>
         struct quote
         {
@@ -226,6 +324,7 @@ namespace meta
 
         /// Turn a class template or alias template \p F taking literals of
         /// type \p T into a Metafunction Class.
+        /// \ingroup composition
         template <typename T, template <T...> class F>
         struct quote_i
         {
@@ -248,6 +347,7 @@ namespace meta
         };
 
         /// Turn a metafunction \p C into a Metafunction Class.
+        /// \ingroup composition
         template <template <typename...> class C>
         struct quote_trait
         {
@@ -257,6 +357,7 @@ namespace meta
 
         /// Turn a metafunction \p C taking literals of type \p T into a
         /// Metafunction Class.
+        /// \ingroup composition
         template <typename T, template <T...> class C>
         struct quote_trait_i
         {
@@ -266,6 +367,7 @@ namespace meta
 
         /// Compose the Metafunction Classes \p Fs in the parameter pack
         /// \p Ts.
+        /// \ingroup composition
         template <typename... Fs>
         struct compose
         {
@@ -287,12 +389,15 @@ namespace meta
 
         namespace lazy
         {
+            /// \sa 'meta::compose'
+            /// \ingroup lazy_composition
             template <typename... Fns>
             using compose = defer<compose, Fns...>;
         }
 
         /// A Metafunction Class that partially applies the Metafunction Class
         /// \p F by binding the arguments \p Ts to the \e front of \p F.
+        /// \ingroup composition
         template <typename F, typename... Ts>
         struct bind_front
         {
@@ -302,6 +407,7 @@ namespace meta
 
         /// A Metafunction Class that partially applies the Metafunction Class
         /// \p F by binding the arguments \p Us to the \e back of \p F.
+        /// \ingroup composition
         template <typename F, typename... Us>
         struct bind_back
         {
@@ -311,35 +417,47 @@ namespace meta
 
         namespace lazy
         {
+            /// \sa 'meta::bind_front'
+            /// \ingroup lazy_composition
             template <typename Fn, typename... Ts>
             using bind_front = defer<bind_front, Fn, Ts...>;
 
+            /// \sa 'meta::bind_back'
+            /// \ingroup lazy_composition
             template <typename Fn, typename... Ts>
             using bind_back = defer<bind_back, Fn, Ts...>;
         }
 
-        /// A metafunction that unpacks the types in the type list
-        /// \p List into the Metafunction Class \p F.
-        template <typename F, typename List>
-        struct apply_list_impl
+        ////////////////////////////////////////////////////////////////////////////
+        // apply_list
+        ///\cond
+        namespace detail
         {
-        };
+            /// A metafunction that unpacks the types in the type list
+            /// \p List into the Metafunction Class \p F.
+            template <typename F, typename List>
+            struct apply_list_
+            {
+            };
 
-        template <typename F, template <typename...> class T, typename... Ts>
-        struct apply_list_impl<F, T<Ts...>> : lazy::apply<F, Ts...>
-        {
-        };
+            template <typename F, template <typename...> class T, typename... Ts>
+            struct apply_list_<F, T<Ts...>> : lazy::apply<F, Ts...>
+            {
+            };
 
-        template <typename F, typename T, T... Is>
-        struct apply_list_impl<F, integer_sequence<T, Is...>>
-            : lazy::apply<F, std::integral_constant<T, Is>...>
-        {
-        };
+            template <typename F, typename T, T... Is>
+            struct apply_list_<F, integer_sequence<T, Is...>>
+                : lazy::apply<F, std::integral_constant<T, Is>...>
+            {
+            };
+        } // namespace detail
+        /// \endcond
 
         /// Applies the Metafunction Class \p C using the types in
         /// the type list \p List as arguments.
+        /// \ingroup invocation
         template <typename C, typename List>
-        using apply_list = eval<apply_list_impl<C, List>>;
+        using apply_list = eval<detail::apply_list_<C, List>>;
 
         namespace lazy
         {
@@ -350,25 +468,31 @@ namespace meta
         /// A Metafunction Class that takes a bunch of arguments, bundles them
         /// into a type list, and then calls the Metafunction Class \p F with the
         /// type list \p Q.
+        /// \ingroup composition
         template <typename F, typename Q = quote<list>>
         using curry = compose<F, Q>;
 
         /// A Metafunction Class that takes a type list, unpacks the types, and
         /// then calls the Metafunction Class \p F with the types.
+        /// \ingroup composition
         template <typename F>
         using uncurry = bind_front<quote<apply_list>, F>;
 
         namespace lazy
         {
+            /// \sa 'meta::curry'
+            /// \ingroup lazy_composition
             template <typename F, typename Q = quote<list>>
             using curry = defer<curry, F, Q>;
 
+            /// \sa 'meta::uncurry'
+            /// \ingroup lazy_composition
             template <typename F>
             using uncurry = defer<uncurry, F>;
         }
 
-        /// A Metafunction Class that reverses the order of the first two
-        /// arguments.
+        /// A Metafunction Class that reverses the order of the first two arguments.
+        /// \ingroup composition
         template <typename F>
         struct flip
         {
@@ -389,6 +513,8 @@ namespace meta
 
         namespace lazy
         {
+            /// \sa 'meta::flip'
+            /// \ingroup lazy_composition
             template <typename... Ts>
             using flip = defer<flip, Ts...>;
         }
@@ -416,15 +542,19 @@ namespace meta
           /// \endcond
 
         /// Select one type or another depending on a compile-time Boolean.
+        /// \ingroup logical
         template <typename... Args>
         using if_ = eval<detail::_if_<Args...>>;
 
         /// Select one type or another depending on a compile-time Boolean.
+        /// \ingroup logical
         template <bool If, typename... Args>
         using if_c = eval<detail::_if_<bool_<If>, Args...>>;
 
         namespace lazy
         {
+            /// \sa 'meta::if_'
+            /// \ingroup lazy_logical
             template <typename... Args>
             using if_ = defer<if_, Args...>;
         }
@@ -462,51 +592,65 @@ namespace meta
         /// \endcond
 
         /// Logically negate the Boolean parameter
+        /// \ingroup logical
         template <bool Bool>
         using not_c = bool_<!Bool>;
 
         /// Logically negate the integral constant-wrapped Boolean parameter.
+        /// \ingroup logical
         template <typename Bool>
         using not_ = not_c<Bool::type::value>;
 
         /// Logically and together all the Boolean parameters
+        /// \ingroup logical
         template <bool... Bools>
         using and_c = std::is_same<integer_sequence<bool, Bools...>,
                                    integer_sequence<bool, (Bools || true)...>>;
 
         /// Logically and together all the integral constant-wrapped Boolean
         /// parameters, \e without doing short-circuiting.
+        /// \ingroup logical
         template <typename... Bools>
         using fast_and = and_c<Bools::type::value...>;
 
         /// Logically and together all the integral constant-wrapped Boolean
         /// parameters, \e with short-circuiting.
+        /// \ingroup logical
         template <typename... Bools>
         using and_ = eval<detail::_and_<Bools...>>;
 
         /// Logically or together all the Boolean parameters
+        /// \ingroup logical
         template <bool... Bools>
         using or_c = not_<std::is_same<integer_sequence<bool, Bools...>,
                                        integer_sequence<bool, (Bools && false)...>>>;
 
         /// Logically or together all the integral constant-wrapped Boolean
         /// parameters, \e without doing short-circuiting.
+        /// \ingroup logical
         template <typename... Bools>
         using fast_or = or_c<Bools::type::value...>;
 
         /// Logically or together all the integral constant-wrapped Boolean
         /// parameters, \e with short-circuiting.
+        /// \ingroup logical
         template <typename... Bools>
         using or_ = eval<detail::_or_<Bools...>>;
 
         namespace lazy
         {
+            /// \sa 'meta::and_'
+            /// \ingroup lazy_logical
             template <typename... Bools>
             using and_ = defer<and_, Bools...>;
 
+            /// \sa 'meta::or_'
+            /// \ingroup lazy_logical
             template <typename... Bools>
             using or_ = defer<or_, Bools...>;
 
+            /// \sa 'meta::not_'
+            /// \ingroup lazy_logical
             template <typename Bool>
             using not_ = defer<not_, Bool>;
         }
@@ -514,6 +658,7 @@ namespace meta
         ///////////////////////////////////////////////////////////////////////////////////////////
         // list
         /// A list of types.
+        /// \ingroup list
         template <typename... Ts>
         struct list
         {
@@ -525,13 +670,15 @@ namespace meta
         ///////////////////////////////////////////////////////////////////////////////////////////
         // size
         /// An integral constant wrapper that is the size of the \c meta::list
-        /// \p
-        /// List.
+        /// \p List.
+        /// \ingroup list
         template <typename List>
         using size = meta::size_t<List::size()>;
 
         namespace lazy
         {
+            /// \sa 'meta::size'
+            /// \ingroup lazy_list
             template <typename List>
             using size = defer<size, List>;
         }
@@ -582,11 +729,14 @@ namespace meta
         /// \pre The parameters must all be instantiations of \c meta::list.
         /// \par Complexity
         /// \f$ O(L) \f$ where \f$ L \f$ is the number of lists in the list of lists.
+        /// \ingroup transformation
         template <typename... Lists>
         using concat = eval<detail::concat_<Lists...>>;
 
         namespace lazy
         {
+            /// \sa 'meta::concat'
+            /// \ingroup lazy_transformation
             template <typename... Lists>
             using concat = defer<concat, Lists...>;
         }
@@ -597,11 +747,14 @@ namespace meta
         /// \par Complexity
         /// \f$ O(L) \f$ where \f$ L \f$ is the number of lists in the list of
         /// lists.
+        /// \ingroup transformation
         template <typename ListOfLists>
         using join = apply_list<quote<concat>, ListOfLists>;
 
         namespace lazy
         {
+            /// \sa 'meta::join'
+            /// \ingroup lazy_transformation
             template <typename ListOfLists>
             using join = defer<join, ListOfLists>;
         }
@@ -635,17 +788,21 @@ namespace meta
         /// Generate `list<T,T,T...T>` of size \p N arguments.
         /// \par Complexity
         /// \f$ O(log N) \f$.
+        /// \ingroup list
         template <typename N, typename T = void>
         using repeat_n = eval<detail::repeat_n_c_<N::type::value, T>>;
 
         /// Generate `list<T,T,T...T>` of size \p N arguments.
         /// \par Complexity
         /// \f$ O(log N) \f$.
+        /// \ingroup list
         template <std::size_t N, typename T = void>
         using repeat_n_c = eval<detail::repeat_n_c_<N, T>>;
 
         namespace lazy
         {
+            /// \sa 'meta::repeat_n'
+            /// \ingroup lazy_list
             template <typename N, typename T = void>
             using repeat_n = defer<repeat_n, N, T>;
         }
@@ -686,17 +843,21 @@ namespace meta
         /// Return the \p N th element in the \c meta::list \p List.
         /// \par Complexity
         /// Amortized \f$ O(1) \f$.
+        /// \ingroup list
         template <typename N, typename List>
         using list_element = eval<detail::list_element_<N, List>>;
 
         /// Return the \p N th element in the \c meta::list \p List.
         /// \par Complexity
         /// Amortized \f$ O(1) \f$.
+        /// \ingroup list
         template <std::size_t N, typename List>
         using list_element_c = list_element<meta::size_t<N>, List>;
 
         namespace lazy
         {
+            /// \sa 'meta::list_element'
+            /// \ingroup lazy_list
             template <typename N, typename List>
             using list_element = defer<list_element, N, List>;
         }
@@ -739,17 +900,21 @@ namespace meta
         /// Return a new \c meta::list by removing the first \p N elements from \p List.
         /// \par Complexity
         /// \f$ O(1) \f$.
+        /// \ingroup transformation
         template <typename N, typename List>
         using drop = eval<detail::drop_<N, List>>;
 
         /// Return a new \c meta::list by removing the first \p N elements from \p List.
         /// \par Complexity
         /// \f$ O(1) \f$.
+        /// \ingroup transformation
         template <std::size_t N, typename List>
         using drop_c = eval<detail::drop_<meta::size_t<N>, List>>;
 
         namespace lazy
         {
+            /// \sa 'meta::drop'
+            /// \ingroup lazy_transformation
             template <typename N, typename List>
             using drop = defer<drop, N, List>;
         }
@@ -775,11 +940,14 @@ namespace meta
         /// Return the first element in \c meta::list \p List.
         /// \par Complexity
         /// \f$ O(1) \f$.
+        /// \ingroup list
         template <typename List>
         using front = eval<detail::front_<List>>;
 
         namespace lazy
         {
+            /// \sa 'meta::front'
+            /// \ingroup lazy_list
             template <typename List>
             using front = defer<front, List>;
         }
@@ -805,11 +973,14 @@ namespace meta
         /// Return the last element in \c meta::list \p List.
         /// \par Complexity
         /// Amortized \f$ O(1) \f$.
+        /// \ingroup list
         template <typename List>
         using back = eval<detail::back_<List>>;
 
         namespace lazy
         {
+            /// \sa 'meta::back'
+            /// \ingroup lazy_list
             template <typename List>
             using back = defer<back, List>;
         }
@@ -835,11 +1006,14 @@ namespace meta
         /// Return a new \c meta::list by adding the element \c T to the front of \p List.
         /// \par Complexity
         /// \f$ O(1) \f$.
+        /// \ingroup transformation
         template <typename List, typename T>
         using push_front = eval<detail::push_front_<List, T>>;
 
         namespace lazy
         {
+            /// \sa 'meta::push_front'
+            /// \ingroup lazy_transformation
             template <typename List, typename T>
             using push_front = defer<push_front, List, T>;
         }
@@ -865,11 +1039,14 @@ namespace meta
         /// Return a new \c meta::list by removing the first element from the front of \p List.
         /// \par Complexity
         /// \f$ O(1) \f$.
+        /// \ingroup transformation
         template <typename List>
         using pop_front = eval<detail::pop_front_<List>>;
 
         namespace lazy
         {
+            /// \sa 'meta::pop_front'
+            /// \ingroup lazy_transformation
             template <typename List>
             using pop_front = defer<pop_front, List>;
         }
@@ -897,11 +1074,14 @@ namespace meta
         /// \f$ O(1) \f$.
         /// \note pop_back not provided because it cannot be made to meet the
         /// complexity guarantees one would expect.
+        /// \ingroup transformation
         template <typename List, typename T>
         using push_back = eval<detail::push_back_<List, T>>;
 
         namespace lazy
         {
+            /// \sa 'meta::push_back'
+            /// \ingroup lazy_transformation
             template <typename List, typename T>
             using push_back = defer<push_back, List, T>;
         }
@@ -910,11 +1090,16 @@ namespace meta
         // empty
         /// An Boolean integral constant wrapper around \c true if \p List is an
         /// empty type list; \c false, otherwise.
+        /// \par Complexity
+        /// \f$ O(1) \f$.
+        /// \ingroup list
         template <typename List>
         using empty = bool_<0 == size<List>::type::value>;
 
         namespace lazy
         {
+            /// \sa 'meta::empty'
+            /// \ingroup lazy_list
             template <typename List>
             using empty = defer<empty, List>;
         }
@@ -950,11 +1135,14 @@ namespace meta
 
         /// Return the tail of the list \p List starting at the first occurrence of
         /// \p T, if any such element exists; the empty list, otherwise.
+        /// \ingroup query
         template <typename List, typename T>
         using find = eval<detail::find_<List, T>>;
 
         namespace lazy
         {
+            /// \sa 'meta::find'
+            /// \ingroup lazy_query
             template <typename List, typename T>
             using find = defer<find, List, T>;
         }
@@ -989,11 +1177,14 @@ namespace meta
 
         /// Return the tail of the list \p List starting at the last occurrence
         /// of \p T, if any such element exists; the empty list, otherwise.
+        /// \ingroup query
         template <typename List, typename T>
         using rfind = eval<detail::rfind_<List, T>>;
 
         namespace lazy
         {
+            /// \sa 'meta::rfind'
+            /// \ingroup lazy_query
             template <typename List, typename T>
             using rfind = defer<rfind, List, T>;
         }
@@ -1025,11 +1216,14 @@ namespace meta
         /// Return the tail of the list \p List starting at the first element `A`
         /// such that `apply<Fun, A>::%value` is \c true, if any such element
         /// exists; the empty list, otherwise.
+        /// \ingroup query
         template <typename List, typename Fun>
         using find_if = eval<detail::find_if_<List, Fun>>;
 
         namespace lazy
         {
+            /// \sa 'meta::find_if'
+            /// \ingroup lazy_query
             template <typename List, typename Fun>
             using find_if = defer<find_if, List, Fun>;
         }
@@ -1061,11 +1255,14 @@ namespace meta
         /// Return the tail of the list \p List starting at the last element `A` such that
         /// `apply<Fun, A>::%value` is \c true, if any such element exists; the empty list,
         /// otherwise.
+        /// \ingroup query
         template <typename List, typename Fun>
         using rfind_if = eval<detail::rfind_if_<List, Fun>>;
 
         namespace lazy
         {
+            /// \sa 'meta::rfind_if'
+            /// \ingroup lazy_query
             template <typename List, typename Fun>
             using rfind_if = defer<rfind_if, List, Fun>;
         }
@@ -1074,11 +1271,14 @@ namespace meta
         // in
         /// A Boolean integral constant wrapper around \c true if there is at least one
         /// occurrence of \p T in \p List.
+        /// \ingroup query
         template <typename List, typename T>
         using in = not_<empty<find<List, T>>>;
 
         namespace lazy
         {
+            /// \sa 'meta::in'
+            /// \ingroup lazy_query
             template <typename List, typename T>
             using in = defer<in, List, T>;
         }
@@ -1112,11 +1312,14 @@ namespace meta
         /// Return a new \c meta::list where all duplicate elements have been removed.
         /// \par Complexity
         /// \f$ O(N^2) \f$.
+        /// \ingroup transformation
         template <typename List>
         using unique = eval<detail::unique_<List, list<>>>;
 
         namespace lazy
         {
+            /// \sa 'meta::unique'
+            /// \ingroup lazy_transformation
             template <typename List>
             using unique = defer<unique, List>;
         }
@@ -1143,11 +1346,14 @@ namespace meta
         /// \p U.
         /// \par Complexity
         /// \f$ O(N) \f$.
+        /// \ingroup transformation
         template <typename List, typename T, typename U>
         using replace = eval<detail::replace_<List, T, U>>;
 
         namespace lazy
         {
+            /// \sa 'meta::replace'
+            /// \ingroup lazy_transformation
             template <typename List, typename T, typename U>
             using replace = defer<replace, T, U>;
         }
@@ -1174,11 +1380,14 @@ namespace meta
         /// `apply<C,A>::%value` is \c true have been replaced with \p U.
         /// \par Complexity
         /// \f$ O(N) \f$.
+        /// \ingroup transformation
         template <typename List, typename C, typename U>
         using replace_if = eval<detail::replace_if_<List, C, U>>;
 
         namespace lazy
         {
+            /// \sa 'meta::replace_if'
+            /// \ingroup lazy_transformation
             template <typename List, typename C, typename U>
             using replace_if = defer<replace_if, C, U>;
         }
@@ -1213,20 +1422,26 @@ namespace meta
         /// `Fun(State_N-1, A_N) -> State_N`.
         /// \par Complexity
         /// \f$ O(N) \f$.
+        /// \ingroup transformation
         template <typename List, typename State, typename Fun>
         using foldl = eval<detail::foldl_<List, State, Fun>>;
 
         /// An alias for `meta::foldl`.
         /// \par Complexity
         /// \f$ O(N) \f$.
+        /// \ingroup transformation
         template <typename List, typename State, typename Fun>
         using accumulate = foldl<List, State, Fun>;
 
         namespace lazy
         {
+            /// \sa 'meta::foldl'
+            /// \ingroup lazy_transformation
             template <typename List, typename State, typename Fun>
             using foldl = defer<foldl, List, State, Fun>;
 
+            /// \sa 'meta::accumulate'
+            /// \ingroup lazy_transformation
             template <typename List, typename State, typename Fun>
             using accumulate = defer<accumulate, List, State, Fun>;
         }
@@ -1258,15 +1473,18 @@ namespace meta
 
         /// Return a new \c meta::list constructed by doing a right fold of the list
         /// \p List using binary Metafunction Class \p Fun and initial state \p State.
-        // That is, the \c State_N for the list element \c A_N is computed by
+        /// That is, the \c State_N for the list element \c A_N is computed by
         /// `Fun(A_N, State_N+1) -> State_N`.
         /// \par Complexity
         /// \f$ O(N) \f$.
+        /// \ingroup transformation
         template <typename List, typename State, typename Fun>
         using foldr = eval<detail::foldr_<List, State, Fun>>;
 
         namespace lazy
         {
+            /// \sa 'meta::foldr'
+            /// \ingroup lazy_transformation
             template <typename List, typename State, typename Fun>
             using foldr = defer<foldr, List, State, Fun>;
         }
@@ -1323,11 +1541,14 @@ namespace meta
         /// lists, pairwise.
         /// \par Complexity
         /// \f$ O(N) \f$.
+        /// \ingroup transformation
         template <typename... Args>
         using transform = eval<detail::transform_<Args...>>;
 
         namespace lazy
         {
+            /// \sa 'meta::transform'
+            /// \ingroup lazy_transformation
             template <typename... Args>
             using transform = defer<transform, Args...>;
         }
@@ -1350,11 +1571,16 @@ namespace meta
         /// Returns a new meta::list where only those elements of \p List A that satisfy the
         /// Metafunction Class \p Predicate such that `apply<Pred,A>::%value` is \c true are
         /// present. That is, those elements that don't satisfy the \p Predicate are "removed".
+        /// \par Complexity
+        /// \f$ O(N) \f$.
+        /// \ingroup transformation
         template <typename List, typename Predicate>
         using filter = meta::foldl<List, meta::list<>, detail::filter_<Predicate>>;
 
         namespace lazy
         {
+            /// \sa 'meta::filter'
+            /// \ingroup lazy_transformation
             template <typename List, typename Predicate>
             using filter = defer<filter, List, Predicate>;
         }
@@ -1394,12 +1620,19 @@ namespace meta
         } // namespace detail
         /// \endcond
 
+        ///\cond
         namespace
         {
+            /// \endcond
+
             /// `for_each(List, UnaryFunction)` calls the \p UnaryFunction for each
             /// argument in the \p List.
+            /// \ingroup runtime
             constexpr auto &&for_each = detail::static_const<detail::for_each_fn>::value;
+
+            /// \cond
         }
+        /// \endcond
 
         ///////////////////////////////////////////////////////////////////////////////////////////
         // zip_with
@@ -1408,6 +1641,7 @@ namespace meta
         /// \par Complexity
         /// \f$ O(N \times M) \f$, where \f$ N \f$ is the size of the outer list, and
         /// \f$ M \f$ is the size of the inner lists.
+        /// \ingroup transformation
         template <typename Fun, typename ListOfLists>
         using zip_with = transform<foldl<ListOfLists, repeat_n<size<front<ListOfLists>>, Fun>,
                                          bind_back<quote<transform>, quote<bind_front>>>,
@@ -1415,6 +1649,8 @@ namespace meta
 
         namespace lazy
         {
+            /// \sa 'meta::zip_with'
+            /// \ingroup lazy_transformation
             template <typename Fun, typename ListOfLists>
             using zip_with = defer<zip_with, Fun, ListOfLists>;
         }
@@ -1426,11 +1662,14 @@ namespace meta
         /// \par Complexity
         /// \f$ O(N \times M) \f$, where \f$ N \f$ is the size of the outer list, and \f$ M \f$
         /// is the size of the inner lists.
+        /// \ingroup transformation
         template <typename ListOfLists>
         using zip = zip_with<quote<list>, ListOfLists>;
 
         namespace lazy
         {
+            /// \sa 'meta::zip'
+            /// \ingroup lazy_transformation
             template <typename ListOfLists>
             using zip = defer<zip, ListOfLists>;
         }
@@ -1454,11 +1693,14 @@ namespace meta
 
         /// Turn a type into an instance of \c meta::list in a way determined by
         /// \c meta::apply_list.
+        /// \ingroup list
         template <typename Sequence>
         using as_list = eval<detail::as_list_<Sequence>>;
 
         namespace lazy
         {
+            /// \sa 'meta::as_list'
+            /// \ingroup lazy_list
             template <typename Sequence>
             using as_list = defer<as_list, Sequence>;
         }
@@ -1468,11 +1710,14 @@ namespace meta
         /// Return a new \c meta::list by reversing the elements in the list \p List.
         /// \par Complexity
         /// \f$ O(N) \f$.
+        /// \ingroup transformation
         template <typename List>
         using reverse = foldr<List, list<>, quote<push_back>>;
 
         namespace lazy
         {
+            /// \sa 'meta::reverse'
+            /// \ingroup lazy_transformation
             template <typename List>
             using reverse = defer<reverse, List>;
         }
@@ -1483,11 +1728,14 @@ namespace meta
         /// is \c true for all elements \c A in \c meta::list \p List; \c false, otherwise.
         /// \par Complexity
         /// \f$ O(N) \f$.
+        /// \ingroup query
         template <typename List, typename F>
         using all_of = empty<find_if<List, compose<quote<not_>, F>>>;
 
         namespace lazy
         {
+            /// \sa 'meta::all_of'
+            /// \ingroup lazy_query
             template <typename List, typename Fn>
             using all_of = defer<all_of, List, Fn>;
         }
@@ -1498,11 +1746,14 @@ namespace meta
         /// \c true for any element \c A in \c meta::list \p List; \c false, otherwise.
         /// \par Complexity
         /// \f$ O(N) \f$.
+        /// \ingroup query
         template <typename List, typename F>
         using any_of = not_<empty<find_if<List, F>>>;
 
         namespace lazy
         {
+            /// \sa 'meta::any_of'
+            /// \ingroup lazy_query
             template <typename List, typename Fn>
             using any_of = defer<any_of, List, Fn>;
         }
@@ -1513,15 +1764,21 @@ namespace meta
         /// \c false for all elements \c A in \c meta::list \p List; \c false, otherwise.
         /// \par Complexity
         /// \f$ O(N) \f$.
+        /// \ingroup query
         template <typename List, typename F>
         using none_of = empty<find_if<List, F>>;
 
         namespace lazy
         {
+            /// \sa 'meta::none_of'
+            /// \ingroup lazy_query
             template <typename List, typename Fn>
             using none_of = defer<none_of, List, Fn>;
         }
 
+        ////////////////////////////////////////////////////////////////////////////
+        // lambda_
+        /// \cond
         namespace detail
         {
             template <int, typename... As>
@@ -1568,6 +1825,7 @@ namespace meta
                 using apply = eval<if_c<sizeof...(Ts) == arity, impl<F, list<Ts..., void>>>>;
             };
         }
+        /// \endcond
 
         ///////////////////////////////////////////////////////////////////////////////////////////
         // lambda
@@ -1577,19 +1835,25 @@ namespace meta
         /// using P = apply<L, int, short>;
         /// static_assert(std::is_same<P, std::pair<short, std::pair<int, int>>>::value, "");
         /// \endcode
+        /// \ingroup metafunction
         template <typename... Ts>
         using lambda = detail::lambda_<0, Ts...>;
 
         // Some argument placeholders for use in \c lambda expressions.
-        struct _a;
-        struct _b;
-        struct _c;
-        struct _d;
-        struct _e;
-        struct _f;
-        struct _g;
-        struct _h;
-        struct _i;
+        /// \ingroup metafunction
+        namespace placeholders
+        {
+            struct _a;
+            struct _b;
+            struct _c;
+            struct _d;
+            struct _e;
+            struct _f;
+            struct _g;
+            struct _h;
+            struct _i;
+        } // namespace placeholders
+        using namespace placeholders;
 
         ///////////////////////////////////////////////////////////////////////////////////////////
         // cartesian_product
@@ -1616,12 +1880,15 @@ namespace meta
         /// \par Complexity
         /// \f$ O(N \times M) \f$, where \f$ N \f$ is the size of the outer list, and
         /// \f$ M \f$ is the size of the inner lists.
+        /// \ingroup transformation
         template <typename ListOfLists>
         using cartesian_product =
             foldr<ListOfLists, list<list<>>, quote_trait<detail::cartesian_product_fn>>;
 
         namespace lazy
         {
+            /// \sa 'meta::cartesian_product'
+            /// \ingroup lazy_transformation
             template <typename ListOfLists>
             using cartesian_product = defer<cartesian_product, ListOfLists>;
         }
@@ -1640,158 +1907,212 @@ namespace meta
         // Math operations
         /// An integral constant wrapper around the result of adding the two wrapped integers
         /// \c T::type::value and \c U::type::value.
+        /// \ingroup math
         template <typename T, typename U>
         using plus = std::integral_constant<decltype(T::type::value + U::type::value),
                                             T::type::value + U::type::value>;
 
         /// An integral constant wrapper around the result of subtracting the two wrapped integers
         /// \c T::type::value and \c U::type::value.
+        /// \ingroup math
         template <typename T, typename U>
         using minus = std::integral_constant<decltype(T::type::value - U::type::value),
                                              T::type::value - U::type::value>;
 
         /// An integral constant wrapper around the result of multiplying the two wrapped integers
         /// \c T::type::value and \c U::type::value.
+        /// \ingroup math
         template <typename T, typename U>
         using multiplies = std::integral_constant<decltype(T::type::value *U::type::value),
                                                   T::type::value * U::type::value>;
 
         /// An integral constant wrapper around the result of dividing the two wrapped integers \c
         /// T::type::value and \c U::type::value.
+        /// \ingroup math
         template <typename T, typename U>
         using divides = std::integral_constant<decltype(T::type::value / U::type::value),
                                                T::type::value / U::type::value>;
 
         /// An integral constant wrapper around the remainder of dividing the two wrapped integers
         /// \c T::type::value and \c U::type::value.
+        /// \ingroup math
         template <typename T>
         using negate = std::integral_constant<decltype(-T::type::value), -T::type::value>;
 
         /// An integral constant wrapper around the remainder of dividing the two wrapped integers
         /// \c T::type::value and \c U::type::value.
+        /// \ingroup math
         template <typename T, typename U>
         using modulus = std::integral_constant<decltype(T::type::value % U::type::value),
                                                T::type::value % U::type::value>;
 
         /// A Boolean integral constant wrapper around the result of comparing \c T::type::value
         /// and \c U::type::value for equality.
+        /// \ingroup math
         template <typename T, typename U>
         using equal_to = bool_<T::type::value == U::type::value>;
 
         /// A Boolean integral constant wrapper around the result of comparing \c T::type::value and
         /// \c U::type::value for inequality.
+        /// \ingroup math
         template <typename T, typename U>
         using not_equal_to = bool_<T::type::value != U::type::value>;
 
         /// A Boolean integral constant wrapper around \c true if \c T::type::value is greater than
         /// \c U::type::value; \c false, otherwise.
+        /// \ingroup math
         template <typename T, typename U>
         using greater = bool_<(T::type::value > U::type::value)>;
 
         /// A Boolean integral constant wrapper around \c true if \c T::type::value is less than \c
         /// U::type::value; \c false, otherwise.
+        /// \ingroup math
         template <typename T, typename U>
         using less = bool_<(T::type::value < U::type::value)>;
 
         /// A Boolean integral constant wrapper around \c true if \c T::type::value is greater than
         /// or equal to \c U::type::value; \c false, otherwise.
+        /// \ingroup math
         template <typename T, typename U>
         using greater_equal = bool_<(T::type::value >= U::type::value)>;
 
         /// A Boolean integral constant wrapper around \c true if \c T::type::value is less than or
         /// equal to \c U::type::value; \c false, otherwise.
+        /// \ingroup math
         template <typename T, typename U>
         using less_equal = bool_<(T::type::value <= U::type::value)>;
 
         /// An integral constant wrapper around the result of bitwise-and'ing the two wrapped
         /// integers \c T::type::value and \c U::type::value.
+        /// \ingroup math
         template <typename T, typename U>
         using bit_and = std::integral_constant<decltype(T::type::value &U::type::value),
                                                T::type::value & U::type::value>;
 
         /// An integral constant wrapper around the result of bitwise-or'ing the two wrapped
         /// integers \c T::type::value and \c U::type::value.
+        /// \ingroup math
         template <typename T, typename U>
         using bit_or = std::integral_constant<decltype(T::type::value | U::type::value),
                                               T::type::value | U::type::value>;
 
         /// An integral constant wrapper around the result of bitwise-exclusive-or'ing the two
         /// wrapped integers \c T::type::value and \c U::type::value.
+        /// \ingroup math
         template <typename T, typename U>
         using bit_xor = std::integral_constant<decltype(T::type::value ^ U::type::value),
                                                T::type::value ^ U::type::value>;
 
         /// An integral constant wrapper around the result of bitwise-complimenting the wrapped
         /// integer \c T::type::value.
+        /// \ingroup math
         template <typename T>
         using bit_not = std::integral_constant<decltype(~T::type::value), ~T::type::value>;
 
         namespace lazy
         {
+            /// \sa 'meta::plus'
+            /// \ingroup lazy_math
             template <typename T, typename U>
             using plus = defer<plus, T, U>;
 
+            /// \sa 'meta::minus'
+            /// \ingroup lazy_math
             template <typename T, typename U>
             using minus = defer<minus, T, U>;
 
+            /// \sa 'meta::multiplies'
+            /// \ingroup lazy_math
             template <typename T, typename U>
             using multiplies = defer<multiplies, T, U>;
 
+            /// \sa 'meta::divides'
+            /// \ingroup lazy_math
             template <typename T, typename U>
             using divides = defer<divides, T, U>;
 
+            /// \sa 'meta::negate'
+            /// \ingroup lazy_math
             template <typename T>
             using negate = defer<negate, T>;
 
+            /// \sa 'meta::modulus'
+            /// \ingroup lazy_math
             template <typename T, typename U>
             using modulus = defer<modulus, T, U>;
 
+            /// \sa 'meta::equal_to'
+            /// \ingroup lazy_math
             template <typename T, typename U>
             using equal_to = defer<equal_to, T, U>;
 
+            /// \sa 'meta::not_equal_t'
+            /// \ingroup lazy_math
             template <typename T, typename U>
             using not_equal_to = defer<not_equal_to, T, U>;
 
+            /// \sa 'meta::greater'
+            /// \ingroup lazy_math
             template <typename T, typename U>
             using greater = defer<greater, T, U>;
 
+            /// \sa 'meta::less'
+            /// \ingroup lazy_math
             template <typename T, typename U>
             using less = defer<less, T, U>;
 
+            /// \sa 'meta::greater_equal'
+            /// \ingroup lazy_math
             template <typename T, typename U>
             using greater_equal = defer<greater_equal, T, U>;
 
+            /// \sa 'meta::less_equal'
+            /// \ingroup lazy_math
             template <typename T, typename U>
             using less_equal = defer<less_equal, T, U>;
 
+            /// \sa 'meta::bit_and'
+            /// \ingroup lazy_math
             template <typename T, typename U>
             using bit_and = defer<bit_and, T, U>;
 
+            /// \sa 'meta::bit_or'
+            /// \ingroup lazy_math
             template <typename T, typename U>
             using bit_or = defer<bit_or, T, U>;
 
+            /// \sa 'meta::bit_xor'
+            /// \ingroup lazy_math
             template <typename T, typename U>
             using bit_xor = defer<bit_xor, T, U>;
 
+            /// \sa 'meta::bit_not'
+            /// \ingroup lazy_math
             template <typename T>
             using bit_not = defer<bit_not, T>;
         }
 
         /// An integral constant wrapper around the minimum of \c T::type::value
         /// and \c U::type::value
+        /// \ingroup math
         template <typename T, typename U>
         using min = meta::if_<meta::less<U, T>, U, T>;
 
         /// An integral constant wrapper around the maximum of \c T::type::value
         /// and \c U::type::value
+        /// \ingroup math
         template <typename T, typename U>
         using max = meta::if_<meta::less<U, T>, T, U>;
 
         namespace lazy
         {
+            /// \sa 'meta::max'
+            /// \ingroup lazy_math
             template <typename T, typename U>
             using max = defer<max, T, U>;
 
+            /// \sa 'meta::min'
+            /// \ingroup lazy_math
             template <typename T, typename U>
             using min = defer<min, T, U>;
         }
@@ -1800,6 +2121,7 @@ namespace meta
         // integer_sequence
 
         /// A container for a sequence of compile-time integer constants.
+        /// \ingroup integral
         template <typename T, T... Is>
         struct integer_sequence
         {
@@ -1848,24 +2170,195 @@ namespace meta
         /// Generate \c integer_sequence containing integer constants [0,1,2,...,N-1].
         /// \par Complexity
         /// \f$ O(log(N)) \f$.
+        /// \ingroup integral
         template <typename T, T N>
         using make_integer_sequence = meta::eval<detail::make_integer_sequence_<T, (std::size_t)N>>;
 
         /// A container for a sequence of compile-time integer constants of type
         /// \c std::size_t
+        /// \ingroup integral
         template <std::size_t... Is>
         using index_sequence = integer_sequence<std::size_t, Is...>;
 
         /// Generate \c index_sequence containing integer constants [0,1,2,...,N-1].
         /// \par Complexity
         /// \f$ O(log(N)) \f$.
+        /// \ingroup integral
         template <std::size_t N>
         using make_index_sequence = make_integer_sequence<std::size_t, N>;
 
-        ///@}  // group meta
-    }
-} // namespace meta::v1
+        /// \cond
+    } // namespace v1
+    /// \endcond
+} // namespace meta
 
-#include <meta/meta_libcxx_workaround.hpp>
+/// \cond
+#if defined(__clang__) && defined(_LIBCPP_VERSION) && _LIBCPP_VERSION <= 1101
+
+_LIBCPP_BEGIN_NAMESPACE_STD
+template <class>
+class _LIBCPP_TYPE_VIS_ONLY allocator;
+template <class, class>
+struct _LIBCPP_TYPE_VIS_ONLY pair;
+template <class>
+struct _LIBCPP_TYPE_VIS_ONLY hash;
+template <class>
+struct _LIBCPP_TYPE_VIS_ONLY less;
+template <class>
+struct _LIBCPP_TYPE_VIS_ONLY equal_to;
+template <class>
+struct _LIBCPP_TYPE_VIS_ONLY char_traits;
+template <class, class>
+class _LIBCPP_TYPE_VIS_ONLY list;
+template <class, class>
+class _LIBCPP_TYPE_VIS_ONLY forward_list;
+template <class, class>
+class _LIBCPP_TYPE_VIS_ONLY vector;
+template <class, class>
+class _LIBCPP_TYPE_VIS_ONLY deque;
+template <class, class, class>
+class _LIBCPP_TYPE_VIS_ONLY basic_string;
+template <class, class, class, class>
+class _LIBCPP_TYPE_VIS_ONLY map;
+template <class, class, class, class>
+class _LIBCPP_TYPE_VIS_ONLY multimap;
+template <class, class, class>
+class _LIBCPP_TYPE_VIS_ONLY set;
+template <class, class, class>
+class _LIBCPP_TYPE_VIS_ONLY multiset;
+template <class, class, class, class, class>
+class _LIBCPP_TYPE_VIS_ONLY unordered_map;
+template <class, class, class, class, class>
+class _LIBCPP_TYPE_VIS_ONLY unordered_multimap;
+template <class, class, class, class>
+class _LIBCPP_TYPE_VIS_ONLY unordered_set;
+template <class, class, class, class>
+class _LIBCPP_TYPE_VIS_ONLY unordered_multiset;
+template <class, class>
+class _LIBCPP_TYPE_VIS_ONLY queue;
+template <class, class, class>
+class _LIBCPP_TYPE_VIS_ONLY priority_queue;
+template <class, class>
+class _LIBCPP_TYPE_VIS_ONLY stack;
+_LIBCPP_END_NAMESPACE_STD
+
+namespace meta
+{
+    inline namespace v1
+    {
+        namespace detail
+        {
+            template <typename T, typename A = std::allocator<T>>
+            using std_list = std::list<T, A>;
+            template <typename T, typename A = std::allocator<T>>
+            using std_forward_list = std::forward_list<T, A>;
+            template <typename T, typename A = std::allocator<T>>
+            using std_vector = std::vector<T, A>;
+            template <typename T, typename A = std::allocator<T>>
+            using std_deque = std::deque<T, A>;
+            template <typename T, typename C = std::char_traits<T>, typename A = std::allocator<T>>
+            using std_basic_string = std::basic_string<T, C, A>;
+            template <typename K, typename V, typename C = std::less<K>,
+                      typename A = std::allocator<std::pair<K const, V>>>
+            using std_map = std::map<K, V, C, A>;
+            template <typename K, typename V, typename C = std::less<K>,
+                      typename A = std::allocator<std::pair<K const, V>>>
+            using std_multimap = std::multimap<K, V, C, A>;
+            template <typename K, typename C = std::less<K>, typename A = std::allocator<K>>
+            using std_set = std::set<K, C, A>;
+            template <typename K, typename C = std::less<K>, typename A = std::allocator<K>>
+            using std_multiset = std::multiset<K, C, A>;
+            template <typename K, typename V, typename H = std::hash<K>,
+                      typename C = std::equal_to<K>,
+                      typename A = std::allocator<std::pair<K const, V>>>
+            using std_unordered_map = std::unordered_map<K, V, H, C, A>;
+            template <typename K, typename V, typename H = std::hash<K>,
+                      typename C = std::equal_to<K>,
+                      typename A = std::allocator<std::pair<K const, V>>>
+            using std_unordered_multimap = std::unordered_multimap<K, V, H, C, A>;
+            template <typename K, typename H = std::hash<K>, typename C = std::equal_to<K>,
+                      typename A = std::allocator<K>>
+            using std_unordered_set = std::unordered_set<K, H, C, A>;
+            template <typename K, typename H = std::hash<K>, typename C = std::equal_to<K>,
+                      typename A = std::allocator<K>>
+            using std_unordered_multiset = std::unordered_multiset<K, H, C, A>;
+            template <typename T, typename C = std_deque<T>>
+            using std_queue = std::queue<T, C>;
+            template <typename T, typename C = std_vector<T>,
+                      class D = std::less<typename C::value_type>>
+            using std_priority_queue = std::priority_queue<T, C, D>;
+            template <typename T, typename C = std_deque<T>>
+            using std_stack = std::stack<T, C>;
+        }
+
+        template <>
+        struct quote<::std::list> : quote<detail::std_list>
+        {
+        };
+        template <>
+        struct quote<::std::deque> : quote<detail::std_deque>
+        {
+        };
+        template <>
+        struct quote<::std::forward_list> : quote<detail::std_forward_list>
+        {
+        };
+        template <>
+        struct quote<::std::vector> : quote<detail::std_vector>
+        {
+        };
+        template <>
+        struct quote<::std::basic_string> : quote<detail::std_basic_string>
+        {
+        };
+        template <>
+        struct quote<::std::map> : quote<detail::std_map>
+        {
+        };
+        template <>
+        struct quote<::std::multimap> : quote<detail::std_multimap>
+        {
+        };
+        template <>
+        struct quote<::std::set> : quote<detail::std_set>
+        {
+        };
+        template <>
+        struct quote<::std::multiset> : quote<detail::std_multiset>
+        {
+        };
+        template <>
+        struct quote<::std::unordered_map> : quote<detail::std_unordered_map>
+        {
+        };
+        template <>
+        struct quote<::std::unordered_multimap> : quote<detail::std_unordered_multimap>
+        {
+        };
+        template <>
+        struct quote<::std::unordered_set> : quote<detail::std_unordered_set>
+        {
+        };
+        template <>
+        struct quote<::std::unordered_multiset> : quote<detail::std_unordered_multiset>
+        {
+        };
+        template <>
+        struct quote<::std::queue> : quote<detail::std_queue>
+        {
+        };
+        template <>
+        struct quote<::std::priority_queue> : quote<detail::std_priority_queue>
+        {
+        };
+        template <>
+        struct quote<::std::stack> : quote<detail::std_stack>
+        {
+        };
+    }
+}
+
+#endif
+/// \endcond
 
 #endif
