+++ readme.md
@@ -2,14 +2,15 @@
 
 [![Build Status](https://travis-ci.org/ericniebler/meta.svg?branch=master)](https://travis-ci.org/ericniebler/meta)
 
-_Meta_ is a C++14 tiny header-only metaprogramming library released under the
+*Meta* is a tiny and header-only C++11 metaprogramming library released under the
 Boost Software License. To compile with meta you just have to:
 
 ```.cpp
 #include <meta/meta.hpp>
 ```
 
-
-
 For a quick start see Eric Niebler's blog post:
 [A tiny metaprogramming library](http://ericniebler.com/2014/11/13/tiny-metaprogramming-library/). (Note: the names in Meta are different from those describe in the blog post, but the overall design remains the same.)
+
+To generate the up-to-date tutorial and documentation run `make doc` in the
+build directory (requires Doxygen).
